修改日志记录变更前后数值，借助脏数据新架构； 日志表LinkId修改为Int64，以支持更大的自增或雪花Id；

+++ XCode/Membership/Interfaces/ILog.cs
@@ -21,7 +21,7 @@ public partial interface ILog
     String Action { get; set; }
 
     /// <summary>链接</summary>
-    Int32 LinkID { get; set; }
+    Int64 LinkID { get; set; }
 
     /// <summary>成功</summary>
     Boolean Success { get; set; }
